Consistent action of conjugation
According to Transform.apply_to_tuple() a potential conjugation is first
applied before the matrix is multiplied. Since the action is from
the left this means that conjugation is on the right.

Thus in Transform.merge() conjugation has to be applied to the entries
of the product of the factors to the right not to the left. This commit
fixes this.